+++ src/main/java/com/newzet/api/auth/business/service/JwtService.java
@@ -9,7 +9,6 @@ import com.newzet.api.auth.business.dto.JwtResponse;
 import com.newzet.api.auth.business.validator.JwtValidator;
 import com.newzet.api.auth.domain.Token;
 import com.newzet.api.auth.exception.TokenBadRequestException;
-import com.newzet.api.auth.infrastructure.TokenRepository;
 
 import lombok.RequiredArgsConstructor;
